# internal/api/testdata/openapi.golden.json
{
  "components": {
    "schemas": {
      "AIAnswer": {
        "properties": {
          "citations": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "grounded": {
            "type": "boolean"
          },
          "sufficient": {
            "type": "boolean"
          },
          "text": {
            "type": "string"
          }
        },
        "required": [
          "text",
          "sufficient"
        ],
        "type": "object"
      },
      "AIQueryRequest": {
        "properties": {
          "limit": {
            "type": "integer"
          },
          "question": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "surfaces": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "surfaces"
        ],
        "type": "object"
      },
      "AIStatus": {
        "properties": {
          "egress": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "endpoint_host": {
            "type": "string"
          },
          "mcp_identity": {
            "type": "string"
          },
          "mcp_write_tools": {
            "type": "boolean"
          },
          "model_configured": {
            "type": "boolean"
          },
          "model_mode": {
            "type": "string"
          },
          "model_name": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "rate_max": {
            "type": "integer"
          },
          "rate_window_seconds": {
            "type": "integer"
          },
          "redaction": {
            "type": "string"
          },
          "residual_refusal_gate": {
            "type": "boolean"
          },
          "runtime": {
            "type": "string"
          }
        },
        "required": [
          "enabled",
          "model_configured",
          "model_mode",
          "egress",
          "redaction",
          "residual_refusal_gate"
        ],
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "trstctl API token",
        "description": "Hashed API token resolved to a tenant-scoped principal with named trstctl permissions.",
        "scheme": "bearer",
        "type": "http"
      },
      "SessionCookie": {
        "description": "Verified browser session from OIDC, SAML, or LDAP; mutating requests also require the double-submit CSRF token.",
        "in": "cookie",
        "name": "trstctl_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Resource-oriented REST API for trstctl. Mutations require an Idempotency-Key; errors are RFC 7807 problem+json; lists use cursor pagination.",
    "title": "trstctl API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ai/query": {
      "post": {
        "operationId": "aiQuery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIAnswer"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Answer a typed semantic/NL query over the tenant's data (read-only, grounded)",
        "x-trstctl-permission": "graph:read"
      }
    },
...
    "/api/v1/ai/status": {
      "get": {
        "operationId": "aiStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIStatus"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Report the served AI runtime/model posture without leaking model endpoint secrets",
        "x-trstctl-permission": "graph:read"
      }
    },
...
  }
}